make publish status its own component
